voice-agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
websockets==13.1
langchain==0.3.7
langchain-openai==0.2.9
openai==1.54.0
python-dotenv==1.0.1
supabase==2.10.0
python-multipart==0.0.12
